+++ packages/enoki-connect/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @mysten/enoki-connect
 
+## 0.0.5
+
+### Patch Changes
+
+- 0bb98c5: update metadata api
+- 8126945: add devnet to supported networks
+- Updated dependencies [afbbb80]
+  - @mysten/wallet-standard@0.15.0
+
+## 0.0.4
+
+### Patch Changes
+
+- ff4193f: allow registering multiple enoki-connect wallets
+
+  - removes `registerEnokiConnectWallet`
+  - adds `registerEnokiConnectWallets`
+
+## 0.0.3
+
+### Patch Changes
+
+- @mysten/sui@1.29.1
+- @mysten/wallet-standard@0.14.9
+
+## 0.0.2
+
+### Patch Changes
+
+- Updated dependencies [7d66a32]
+- Updated dependencies [eb91fba]
+- Updated dependencies [19a8045]
+  - @mysten/sui@1.29.0
+  - @mysten/wallet-standard@0.14.8
+
 ## 0.0.1
 
 ### Patch Changes