+++ packages/enoki/CHANGELOG.md
@@ -1,5 +1,38 @@
 # @mysten/enoki
 
+## 0.6.11
+
+### Patch Changes
+
+- Updated dependencies [afbbb80]
+  - @mysten/wallet-standard@0.15.0
+
+## 0.6.10
+
+### Patch Changes
+
+- @mysten/sui@1.29.1
+- @mysten/signers@0.2.8
+- @mysten/wallet-standard@0.14.9
+
+## 0.6.9
+
+### Patch Changes
+
+- Updated dependencies [8356318]
+  - @mysten/signers@0.2.7
+
+## 0.6.8
+
+### Patch Changes
+
+- Updated dependencies [7d66a32]
+- Updated dependencies [eb91fba]
+- Updated dependencies [19a8045]
+  - @mysten/sui@1.29.0
+  - @mysten/signers@0.2.6
+  - @mysten/wallet-standard@0.14.8
+
 ## 0.6.7
 
 ### Patch Changes