+++ packages/mvr-static/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @mysten/mvr-static
 
+## 0.1.10
+
+### Patch Changes
+
+- @mysten/sui@1.29.1
+
+## 0.1.9
+
+### Patch Changes
+
+- Updated dependencies [7d66a32]
+- Updated dependencies [eb91fba]
+- Updated dependencies [19a8045]
+  - @mysten/sui@1.29.0
+
 ## 0.1.8
 
 ### Patch Changes