+++ packages/seal/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @mysten/seal
 
+## 0.4.3
+
+### Patch Changes
+
+- d3f0e8d: Export SessionKeyType
+
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [7e1c525]
+  - @mysten/bcs@1.6.1
+  - @mysten/sui@1.29.1
+
+## 0.4.1
+
+### Patch Changes
+
+- 81f406a: Add more details to InvalidPTB error
+- 7f108cb: feat(seal): add import export to session key
+- Updated dependencies [7d66a32]
+- Updated dependencies [eb91fba]
+- Updated dependencies [19a8045]
+  - @mysten/sui@1.29.0
+
 ## 0.4.0
 
 ### Minor Changes