+++ packages/slush-wallet/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @mysten/slush-wallet
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [afbbb80]
+  - @mysten/wallet-standard@0.15.0
+
+## 0.1.2
+
+### Patch Changes
+
+- 3f87e73: Fix broken message signing for zkLogin accounts in the web version of Slush Wallet
+  - @mysten/sui@1.29.1
+  - @mysten/wallet-standard@0.14.9
+
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [7d66a32]
+- Updated dependencies [eb91fba]
+- Updated dependencies [19a8045]
+  - @mysten/sui@1.29.0
+  - @mysten/wallet-standard@0.14.8
+
 ## 0.1.0
 
 ### Minor Changes